INC-105 fix: imports, refactor user settings page, added loading profile from server and set it to the user general information form
