+++ src/app/index.ts
@@ -1,8 +1,6 @@
 export * from './assets'
-export * from './constants'
 export * from './helpers'
 export * from './hooks'
 export * from './locales'
-export * from './utils'
 export * from './types'
 export * from './data'
