--- src/app/services/profile/index.ts
@@ -1,2 +1,0 @@
-export * from './profile.slice'
-export * from './profile.api.types'
