+++ src/app/store/store.ts
@@ -5,7 +5,7 @@ import { authReducer } from '@/app/services/auth/auth.slice'
 import { commonApi } from '@/app/services/common/common.api'
 import { locationApi } from '@/app/services/countries/countries.api'
 import { googleApi } from '@/app/services/google/google.api'
-import { profileSlice } from '@/app/services/profile'
+import { profileSlice } from '@/app/services/profile/profile.slice'
 
 export const store = configureStore({
   reducer: {
