+++ src/modules/create-new-password-form/create-new-password-confirmation/create-new-password-confirmation.tsx
@@ -5,7 +5,8 @@ import { useRouter } from 'next/router'
 
 import s from './create-new-password-confirmation.module.scss'
 
-import { authNavigationUrls, useTranslation } from '@/app'
+import { useTranslation } from '@/app'
+import { authNavigationUrls } from '@/app/constants'
 import { Typography } from '@/ui'
 
 type ConfirmationProps = {
