+++ src/modules/create-new-password-form/create-new-password.tsx
@@ -2,8 +2,10 @@ import { useRouter } from 'next/router'
 
 import s from './create-new-password.module.scss'
 
-import { authNavigationUrls, ErrorWithData, showError, useTranslation } from '@/app'
+import { ErrorWithData, useTranslation } from '@/app'
+import { authNavigationUrls } from '@/app/constants'
 import { useCreateNewPasswordMutation } from '@/app/services/auth/auth.api'
+import { showError } from '@/app/utils'
 import { useNewPasswordForm } from '@/modules/create-new-password-form/validation-schema'
 import { Button, Card, Loader, TextField, Typography } from '@/ui'
