+++ src/modules/forgot-password-form/forgot-password-form.tsx
@@ -7,7 +7,8 @@ import { useRouter } from 'next/router'
 import s from './forgot-password-form.module.scss'
 import { useForgotPasswordForm } from './validation-schema'
 
-import { authNavigationUrls, useDisclose, useMatchMedia, useTranslation } from '@/app'
+import { useDisclose, useMatchMedia, useTranslation } from '@/app'
+import { authNavigationUrls } from '@/app/constants'
 import { usePasswordRecoveryMutation } from '@/app/services/auth/auth.api'
 import { NotificationModal } from '@/components'
 import { Button, Card, ControlledReCaptcha, Loader, TextField, Typography } from '@/ui'
