+++ src/modules/sign-in-form/sign-in-form.tsx
@@ -7,8 +7,10 @@ import { toast } from 'react-toastify'
 
 import s from './sign-in-form.module.scss'
 
-import { authNavigationUrls, ErrorWithData, showError, useTranslation } from '@/app'
+import { ErrorWithData, useTranslation } from '@/app'
+import { authNavigationUrls } from '@/app/constants'
 import { useSignInMutation } from '@/app/services/auth/auth.api'
+import { showError } from '@/app/utils'
 import { ControlledTextField } from '@/components/text-field-controlled/controlled-text-field'
 import { useSignInForm } from '@/modules/sign-in-form/use-sign-in-form'
 import { Button, Card, GithubButton, GoogleButton, Typography } from '@/ui'
